Aulas

+++ src/br/com/eduardo/banco/modelo/pagamento/Boleto.java
@@ -1,20 +1,22 @@
 package br.com.eduardo.banco.modelo.pagamento;
 
+import java.math.BigDecimal;
+
 import br.com.eduardo.banco.modelo.Pessoa;
 
 public class Boleto implements DocumentoEstornavel {
 
     private Pessoa beneficiario;
-    private double valor;
+    private BigDecimal valor;
     private boolean pago;
 
-    public Boleto(Pessoa beneficiario, double valor) {
+    public Boleto(Pessoa beneficiario, BigDecimal valor) {
         this.beneficiario = beneficiario;
         this.valor = valor;
     }
 
     @Override
-    public double getValorTotal() {
+    public BigDecimal getValorTotal() {
         return valor;
     }
